requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
Flask-SQLAlchemy
flask-jwt-extended
passlib
flask-migrate
pymysql
python-dotenv